feature(revisions): Adds "revisions" extension
Revisions are an extension to the SemVer standard, that allow an
additional value on the end for the core version number, delimited by an
underscore (`_`) character - E.g. `1.2.3_4`.

Revisions are useful in situations where you need to distingush a
software component's version from an upstream change. For example, you
may be packaging a software component for a Linux distribution and you
want to indicate differences in the package itself rather than the
underlying software.